backend/requirements.txt
# Djando + util
Django~=4.0.10
pillow~=9.5.0
sorl-thumbnail~=12.9.0
python-dotenv~=1.0.0

# Model converter
pandas~=2.0.2
openpyxl~=3.1.2

# REST API
djangorestframework~=3.14.0
django-filter~=23.2
drf-extra-fields~=3.5.0
djoser~=2.2.0

# Django productions
gunicorn~=20.1.0
psycopg2-binary~=2.9.6
